requirements.txt
pybtex
python-frontmatter
